// frontend/spotify-infinite-mix/src/model/analysis-response.model.ts
export interface SpotifyAnalysis {
    acousticness: number;
    analysis_url: string;
    danceability: number;
    duration_ms: number;
    energy: number;
    id: string;
    instrumentalness: number;
    key: number;
    liveness: number;
    loudness: number;
    mode: number;
    speechiness: number;
    tempo: number;
    time_signature: number;
    track_href: string;
    type: string;
    uri: string;
    valence: number;
    percent_listened?: number; // Addition to standard model for our y-variable
}